Manager/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0
django-cors-headers==3.10.1
django-phonenumber-field==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
phonenumbers==8.12.39
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
